Use a server action for adding a property and show user feedback
