--- app/api/properties/route.js
@@ -1,103 +1,0 @@
-import connectDB from '@/config/database';
-import Property from '@/models/Property';
-import { getSessionUser } from '@/utils/getSessionUser';
-import cloudinary from '@/config/cloudinary';
-
-// GET /api/properties
-// NOTE: GET handler removed as it is no longer used as we can query the db
-// directly from server components.
-
-export const POST = async (request) => {
-  try {
-    await connectDB();
-
-    const sessionUser = await getSessionUser();
-
-    if (!sessionUser || !sessionUser.userId) {
-      return new Response('User ID is required', { status: 401 });
-    }
-
-    const { userId } = sessionUser;
-
-    const formData = await request.formData();
-
-    // Access all values from amenities and images
-    const amenities = formData.getAll('amenities');
-
-    const images = formData
-      .getAll('images')
-      .filter((image) => image.name !== '');
-
-    // Create propertyData object for database
-    const propertyData = {
-      type: formData.get('type'),
-      name: formData.get('name'),
-      description: formData.get('description'),
-      location: {
-        street: formData.get('location.street'),
-        city: formData.get('location.city'),
-        state: formData.get('location.state'),
-        zipcode: formData.get('location.zipcode'),
-      },
-      // FIX: beds and baths should be of type Number
-      beds: formData.get('beds'),
-      baths: formData.get('baths'),
-      square_feet: formData.get('square_feet'),
-      // FIX: rates in our Property Model are of type 'Number' but from the
-      // formData they will be of Type 'String'
-      amenities,
-      rates: {
-        weekly: formData.get('rates.weekly'),
-        monthly: formData.get('rates.monthly'),
-        nightly: formData.get('rates.nightly.'),
-      },
-      seller_info: {
-        name: formData.get('seller_info.name'),
-        email: formData.get('seller_info.email'),
-        phone: formData.get('seller_info.phone'),
-      },
-      owner: userId,
-    };
-
-    // Upload image(s) to Cloudinary
-    // NOTE: this will be an array of strings, not a array of Promises
-    // So imageUploadPromises has been changed to imageUrls to more
-    // declaratively represent it's type.
-
-    const imageUrls = [];
-
-    for (const imageFile of images) {
-      const imageBuffer = await imageFile.arrayBuffer();
-      const imageArray = Array.from(new Uint8Array(imageBuffer));
-      const imageData = Buffer.from(imageArray);
-
-      // Convert the image data to base64
-      const imageBase64 = imageData.toString('base64');
-
-      // Make request to upload to Cloudinary
-      const result = await cloudinary.uploader.upload(
-        `data:image/png;base64,${imageBase64}`,
-        {
-          folder: 'propertypulse',
-        }
-      );
-
-      imageUrls.push(result.secure_url);
-    }
-
-    // NOTE: here there is no need to await the resolution of
-    // imageUploadPromises as it's not a array of Promises it's an array of
-    // strings, additionally we should not await on every iteration of our loop.
-
-    propertyData.images = imageUrls;
-
-    const newProperty = new Property(propertyData);
-    await newProperty.save();
-
-    return Response.redirect(
-      `${process.env.NEXTAUTH_URL}/properties/${newProperty._id}`
-    );
-  } catch (error) {
-    return new Response('Failed to add property', { status: 500 });
-  }
-};
